Move remaining code in CommandActionHandler to DebugLandsCommand
